1.0.0 updates matc (#3296)
* Removing pointless env utils

* Removing pointless app utils

* Removing pointless helpers utils

* Removing pointless browser utils

* Removing pointless types utils

* Removing pointless media utils

* Removing pointless role utils

* Removing pointless indexes.

* Some reorg.

* Some reorg 2.

* Lemmy UI 1.0.0 update WIP.

* A few more components.

* A few more components.

* Updated dependencies from main

* link:../lemmy-js-client

* Fix most @utils import

* Split MyUserInfo from GetSiteResponse

* Rework i18next and date-fns configuration

date-fns language matches i18n language. SSR creates a new i18n instance
every render. Translations and date-fns locales are loaded ahead of
setting the current language, this allows i18n to be initialized
synchronously.

* Pass myUserInfo around as mandatory property

* updateMyUserInfo

* Fix UnreadCounterService

* Bring back IsoDataOptionalSite

* Vote FederationMode

* HidePost, saved_only, tags, post_ids -> post_id

* admin/userBlockInstance

* ReportCombinedView

* CommunityReportView (TODO: create report)

* Profile PersonContentCombinedView

* Use new image api endpoints

* Remove enable_nsfw and federation_debug

* ModlogCombinedView

* InboxCombinedView

* Mark comment reply/mention as read

* Mark post as read

* Mark page as read with markManyPostAsRead

* SearchCombinedView

* User settings enable_private_messages, auto_mark_fetched_posts_as_read

* can_mod, saved?: string

* actor_id > ap_id

* SortSelect

* Communities page cursor

* Make webpack resolve @serivces @components

* Bring back some date utils

* Fixup allLanguages

* Fixup RateLimitFormProps

* Remove CommentAggregates, PostAggregates, CommunityAggregates, SiteAggregates

* Fix show_score in Settings

* Fix AdminSettings

* Make arguments to @utils/roles mandatory, use can_mod

* Fix PostMention mark as read

* Updating from main.

* A few components.

* Toasts and console warnings for missing i18n keys in development mode

* Stop translations:generate from touching unchanged files

* Cast error messages to NoOptionI18nKeys

* Typed I18NextService.i18n.t()

* Remove some explicit casts to NoOptionI18nKey

* creator_banned, post_tags, tags

* deleteImage -> deleteMedia

* Page limits

* More mandatory MyUserInfo arguments and props

* Collapse blurred nsfw image posts

* VoteButtons, VoteDisplay

* Fix floating point unix time

* Bump lemmy-js-client dependency

* Remove errors from most non-component files

* Fix some TS build errors (there are so many!)

* Get rid of build errors for modlog page

* Comment out remote fetch until we figure out how to handle it with new response

* View votes modal pagination

* Add utility to check if user is banned

* Fix up community sidebar

* Add pagination to admin settings page

* Get rid of typescript errors in emojis-form

* Get AdminSettings and SiteForm (mostly) building in typescript

* Fix TS build errors in Setup

* Fix disallow NSFW part of SiteForm

* Fix search page

* Fix Paginator imports

* Make it so home page can load

* Hide empty moderator list in community sidebar

* Assume post reports are only visible to moderators and the creator

* Image upload confirmation title and description

* Community sidebar, reports

* Use listUsers to list banned users

* VoteShow for show_downvotes

* Translation

* Rename timestamps, use lemmy-js-client@1.0.0-rename-timestamp-to-at.4

* Drop hide_modlog_mod_names

* RemoteFetch using SearchResponse

* SearchResponse includes resolved objects

* Fix Post votes

* lemmy-js-client@1.0.0-rename-rate-limit-columns.1

* Remove creator_community_actions

* Fix Search search_term

* Rate limit renames

* lemmy-js-client@1.0.0-instance-user-blocking.8

* Inbox using listNotifications

* Update modlog pagination again

* Update registration-applications pagination

* Update communities pagination

* Copy package.json, pnpm-lock.yaml from origin/main

* Fixup rate-limit-form

* Fix few linkEvent, bound handler

* Translations

* Update inferno to 9.0.3 (needs webpack and babel target >= es6)

* Cleanup page cursors

* Remove some unused code

* Fix some UI issues.

* i18n

* Fix image upload in create community

* Make error page escapable

* Fix some smaller todos

* Remove mark comment as read outside inbox

* Cleanup mark post as read

* Cleanup mark pm as read

* Remove filtered person-details views, API already filters

* Translations

---------

+++ generate_translations.js
@@ -17,9 +17,14 @@ fs.readdir(translationDir, (_err, files) => {
           data += `\n    ${key}: "${value}",`;
         }
       }
-      data += "\n  },\n};";
+      data += "\n  },\n} as const;";
       const target = outDir + lang + ".ts";
-      fs.writeFileSync(target, data);
+      if (
+        !fs.existsSync(target) ||
+        fs.readFileSync(target).toString() !== data
+      ) {
+        fs.writeFileSync(target, data);
+      }
     } catch (err) {
       console.error(err);
     }
@@ -55,9 +60,17 @@ fs.readFile(`${translationDir}${baseLanguage}.json`, "utf8", (_, fileStr) => {
 
   const indent = "    ";
 
-  const data = `import { i18n } from "i18next";
+  const data = `import "i18next";
+import { en } from "./en";
 
 declare module "i18next" {
+  interface CustomTypeOptions {
+    jsonFormat: "v3"; // no longer supported in ^24.0, (can just be removed after converting v4)
+    // strictKeyChecks: true; would also check that options are provided, needs ^24.2
+    resources: {
+      translation: typeof en.translation
+    };
+  }
   export type NoOptionI18nKeys = 
 ${noOptionKeys.map(key => `${indent}| "${key}"`).join("\n")};
 
@@ -65,44 +78,11 @@ ${noOptionKeys.map(key => `${indent}| "${key}"`).join("\n")};
 ${optionKeys.map(key => `${indent}| "${key}"`).join("\n")};
 
   export type I18nKeys = NoOptionI18nKeys | OptionI18nKeys;
-
-  export type TTypedOptions<TKey extends OptionI18nKeys> =${Array.from(
-    optionMap.entries(),
-  ).reduce(
-    (acc, [key, options]) =>
-      `${acc} TKey extends \"${key}\" ? ${
-        options.reduce((acc, cur) => acc + `${cur}: string | number; `, "{ ") +
-        "}"
-      } :\n${indent}`,
-    "",
-  )} (Record<string, unknown> | string);
-
-  export interface TFunctionTyped {
-    // Translation requires options
-    <
-      TKey extends OptionI18nKeys | OptionI18nKeys[],
-      TResult extends TFunctionResult = string,
-      TInterpolationMap extends TTypedOptions<TKey> = StringMap
-    > (
-      key: TKey,
-      options: TOptions<TInterpolationMap> | string
-    ): TResult;
-
-    // Translation does not require options
-    <
-      TResult extends TFunctionResult = string,
-      TInterpolationMap extends Record<string, unknown> = StringMap
-    > (
-      key: NoOptionI18nKeys | NoOptionI18nKeys[],
-      options?: TOptions<TInterpolationMap> | string
-    ): TResult;
-  }
-
-  export interface i18nTyped extends i18n {
-    t: TFunctionTyped;
-  }
 }
 `;
 
-  fs.writeFileSync(`${outDir}i18next.d.ts`, data);
+  const target = `${outDir}i18next.d.ts`;
+  if (!fs.existsSync(target) || fs.readFileSync(target).toString() !== data) {
+    fs.writeFileSync(target, data);
+  }
 });
